+++ src/main/java/com/accenture/applicationlocationvehicule/repository/entity/Motorcycle.java
@@ -12,7 +12,7 @@ public class Motorcycle {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private int powerInKw;
     private String Transmission;
     private String requiredLicense;
@@ -30,7 +30,7 @@ public class Motorcycle {
     private int displacement;
 
 
-    public Motorcycle(Long id, int powerInKw, String requiredLicense, String transmission, String brand, String model, String color, int numberOfCylinder, int seatHeight, int weight, boolean removedFromFleet, boolean active, int dailyBaseRate, int mileage, String type, int displacement) {
+    public Motorcycle(int id, int powerInKw, String requiredLicense, String transmission, String brand, String model, String color, int numberOfCylinder, int seatHeight, int weight, boolean removedFromFleet, boolean active, int dailyBaseRate, int mileage, String type, int displacement) {
         this.id = id;
         this.powerInKw = powerInKw;
         this.requiredLicense = requiredLicense;
@@ -49,11 +49,11 @@ public class Motorcycle {
         this.displacement = displacement;
     }
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
